+++ src/RE/P/PlayerCharacter.cpp
@@ -259,6 +259,22 @@ namespace RE
 		return func(this);
 	}
 
+	void PlayerCharacter::UpdateVRComfortCheck()
+	{
+		if (REL::Module::IsVR()) {
+			using func_t = decltype(&PlayerCharacter::UpdateVRComfortCheck);
+			static REL::Relocation<func_t> func{ REL::VariantID(0, 0, 0x6B9AA0) };
+			func(this);
+		}
+	}
+
+	void PlayerCharacter::UsePoisonFromInventory(AlchemyItem* a_poison)
+	{
+		using func_t = decltype(&PlayerCharacter::UsePoisonFromInventory);
+		static REL::Relocation<func_t> func{ RELOCATION_ID(39406, 40481) };
+		return func(this, a_poison);
+	}
+
 	bool PlayerCharacter::CenterOnCell_Impl(const char* a_cellName, RE::TESObjectCELL* a_cell)
 	{
 		using func_t = decltype(&PlayerCharacter::CenterOnCell_Impl);